import rust to wasm

+++ rust/src/lib.rs
@@ -1,6 +1,6 @@
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
-pub fn add(a: i32, b: i32) -> i32 {
+pub fn add(a: f32, b: f32) -> f32 {
     a + b
 }
